Added Query.copy and split where into two cases.
So that a query builder can produce a new instance which
does not interfere with the instance it is a copy of, introduced
Query.copy() which does a shallow copy on the query and
of its array contents , but no further.

GraphPatterns can be added in either of the Early or Later
positions. Early patterns appear before later ones in the
generated SPARQL query. This is to correspond to existing
distinctions in DSAPI and SAPI2.