Corrected problem where usePrefix in Settings would deliver bad prefixed names. For example if the prefix p was bound to http://example.com/pre/amble, usePrefix applied to http://example.com/pre/amble/alpha/beta would deliver p:alpha/beta, which is illegal, / not being allowed in the leafname.
Fixed by checking the returned leafname matches the pattern

	"[_A-Za-z0-9]+([-.:][_A-Za-z0-9]+)*[_A-Za-z0-9]$

which only allows letters, digits, and underbars and strict infix
use of colon, hyphen, and dot, eg x.y is allowed but not .xy or
xy. This is (as far as I can see) more restricted than the
Turtle grammar requires.

While Turtle local names allow a whole bunch of non-ascii Unicode characters,
for the moment we wimp out and just disallow them (which just means they
won't be prefix-shortened).